requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
Pillow>=8.0.1,<8.1.0
django-crispy-forms>=1.10.0,<1.11.0
django-markdownx>=3.0.1,<3.1.0
beautifulsoup4>=4.9.3,<4.10.0
django-image-cropping>=1.5.0,<1.6.0
easy-thumbnails>=2.7.1,<2.8.0
drf-yasg>=1.20.0,<1.21.0